requirements.txt
Flask
waitress
python-dotenv
okta
python-ldap
redis
rq
python-json-logger 